% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileOutFilePath.R
\name{compileOutFilePath}
\alias{compileOutFilePath}
\title{Compile output filepathes}
\usage{
my_outfilepathes = compileOutFilePath(input_filepathes, output_subdirectory, prefix=NA, suffix=NA)
}
\arguments{
\item{input_filepathes}{List of input files}

\item{output_subdirectory}{output directory (name for output directory)}

\item{prefix}{Set a new prefix to file-name (default = NA)}

\item{suffix}{Set a new suffix to filename (default = NA)}
}
\description{
Creates output subdirectory for generated data. Directory will be at the same level like input filepath.
It's possible to set different prefix or suffix to filename.
}
\examples{
\dontrun{
compileOutFilePath(input_filepathes = "modis_filled_tiles",
output_subdiretory = "filled", prefix= "tiles", suffix = "filled")
}

}
\author{
A. Santowski & C. Weber
}
