% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileRawTestData.R
\name{compileRawTestData}
\alias{compileRawTestData}
\title{compileRawTestData}
\usage{
compileRawTestData()
}
\description{
Compile raw test data for this package
}
\details{
Compile raw test data
}
\author{
T. Nauss
}
