% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseason.R
\name{deseason}
\alias{deseason}
\title{Remove seasonality from raster time series}
\usage{
deseason(rstack, outFilePath, cycle.window)
}
\arguments{
\item{rstack}{raster stack of scaled, temporal aggregated and filled
(= NA-free) rasters = output of function "fill_gaps_lin" following folder
structure of 00_set_environment.R this would mean: all rasters in path
"path_modis_filled_tiles"}

\item{outFilePath}{see compileOutFilePath()}

\item{cycle.window}{as integer (with "L" in the end of the number, f.ex.: 12L),
12L means monthly anomalies, 24L := half a months, 6L := half a year, etc.}
}
\value{
writes out deseasonalized rasters
}
\description{
Removes seasonality from time series of scaled, temporal
aggregated and filled (= NA-free) rasters (f.ex. NDVI-time-series)
}
\examples{
\dontrun{
deseason(your_raster_stack, outFilePath, 12L)
}

}
\references{
Nauss, T., Detsch, F.
}
\author{
Johannes Schnell, Laura Giese
}
