% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGapsLin.R
\name{fillGapsLin}
\alias{fillGapsLin}
\title{Fill gaps by linear interpolation}
\usage{
fillGapsLin(rst_fn, out_path)
}
\arguments{
\item{rst_fn}{raster stack of scaled and temporal aggregated rasters = output of previous function (3rd function of temporal aggregation),
following folder structure of 00_set_environment.R this would mean: all rasters in path "path_modis_temp_agg_tiles"
rst_fn = stack(list.files(path_modis_temp_agg_tiles, pattern = glob2rx("*.tif"), full.names = TRUE))}

\item{out_path}{path to folder, where output should be stored
following folder structure of 00_set_environment.R this would mean: out_path = paste0(path_modis_filled_tiles)}
}
\value{
writes out rasters with no NA values
}
\description{
Subsets NAs by spacial linear interpolation
}
\examples{
\dontrun{
fillGapsLin(rst_fn, out_path)
}

}
\references{
Nauss, T., Detsch, F.
}
\author{
Johannes Schnell, Laura Giese
}
