% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexTimeAggStack.R
\name{flexTimeAggStack}
\alias{flexTimeAggStack}
\title{flexTimeAggStack}
\usage{
flexTimeAggStack(beginnzeitsp, endzeitsp, dates_path, aggrdata, outfilepath,
  edit = "kap", a = 1, b = 4, c = 5, d = 7)
}
\arguments{
\item{endzeitsp, }{integer, year of end}

\item{dates_path, }{filepath to sorted and listet data with date info}

\item{aggrdata, }{filepath to stored data which shall be temporally aggregated}

\item{outfilepath, }{path to store temporal aggregated tif files}

\item{edit, }{character, shortnote for data [MAX 3 LETTERS]}

\item{a}{integer, first pos of annual date}

\item{b}{integer, last pos of annual date}

\item{c}{integer, first pos of daily date}

\item{d}{integer, last pos of daily date}

\item{beginzeitsp, }{integer, year of start}
}
\value{
geo-tiff data
}
\details{
This function aggregates, matching to destination dates, tiff layers and sums them to a single layer.
dependent on package "raster" & "rgdal"
}
\author{
A. Santowski & L. Giese
#'@examples
\dontrun{
flexTimeAggStack=function(beginnzeitsp,endzeitsp,dates_path,aggrdata,outfilepath,edit="kap")
}
}
