% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkTrend.R
\name{mkTrend}
\alias{mkTrend}
\title{Compute Mann-Kendall trend}
\usage{
trend(file_in, p, prewhitening, method, filename)
}
\arguments{
\item{input}{rasterstack}

\item{p, }{numeric, defaults to 0.001. Significance level to be tested}

\item{prewhitening, logical, }{If TRUE (default), pre-whitening is applied prior to the Mann-Kendall trend test.}

\item{method}{The prewhitening method to apply ("yuepilon", "zhang")}

\item{filename}{name and path of output files}
}
\value{
rst_stau; geo-tiff data
}
\details{
rank correlation coefficient Tau
need raster and gimms package
}
\author{
HammerLe, Kleebaue
}
