% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierCheck.R
\name{outlierCheck}
\alias{outlierCheck}
\title{Additional outlier check}
\usage{
outlierCheck(input, lq, uq)
}
\arguments{
\item{rstack}{Raster stack}

\item{input, }{path where input is}

\item{lq, }{Numeric, lower quantile limit}

\item{uq, }{Numeric, upper quantile limit}

\item{outfile}{vector with outfiles}
}
\value{
geo-tiff data
}
\details{
this is an example function named 'Additional outlier check'
which checks for statistical outliers
}
