% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{qualityCheck}
\alias{qualityCheck}
\title{Check quality of raster data}
\usage{
qualityCheck(rstck_values, rstck_quality, outputfilepathes)
}
\arguments{
\item{rstck_values}{\code{raster::stack} with data values}

\item{rstck_quality}{\code{raster::stack} with ranked quality values}

\item{outputfilepathes}{name and path of output files}
}
\value{
stack of quality checked data values
}
\description{
Checks raster data with a quality layer. Replaces values with bad quality with NA.
}
\examples{
\dontrun{
quality_check(p = c("NDVI.tif$", "reliability.tif$"), path_modis_prj = path_modis_prj, path_modis_quality_checked = path_modis_quality_checked)
}
}
\references{
Nauss, T., Detsch, F.
}
\author{
Johannes Schmidt, Tiziana Koch, Niklas Balzer
}
