% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleRaster.R
\name{scaleRaster}
\alias{scaleRaster}
\alias{Scaling}
\title{scaleRaster}
\usage{
scaleRaster(rstck, scalefac = 10000, incon = NA, outputfilepathes)
}
\arguments{
\item{rstck}{rstck raster stack}

\item{scalefac}{Set scaling factor (default= 10000), factor to divide initial raster values}

\item{incon}{Dealing with rejected values (default = NA)}

\item{outputfilepathes}{Set output file path}
}
\description{
The scaling function takes up a \code{raster::stack} and scales it.
}
\details{
Scale raster data

The function changes the scaling (default = /10000) and rejects inconsistend values (default = NA).
The function writes scaled GeoTiffs and returns also a raster stack.
}
\examples{
\dontrun{
fn <-
scaling(rstck, scalefac = 1000, incon = NA)
}

}
\author{
Santwoski, A. & C. Weber
}
