% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialProjection.R
\name{spatialProjection}
\alias{spatialProjection}
\title{spatial projection}
\usage{
spatialProjection=function(rst_from,prst_to,outfilepath)
dependent on package "raster" and "rgdal"
}
\arguments{
\item{list_rst_from}{charachter string: List of input tif files}

\item{list_prst_to}{charachter string: List of designated projected tif files}

\item{outfilepath}{path where output tif files shall be stored}

\item{areaname}{three letters shortcut for study area}
}
\value{
geo tif data
}
\details{
Projects a tif files into another projection from a given tif file.
}
\examples{
\dontrun{
spatialProjection=function(rst_from,prst_to,outfilepath)
}

}
\author{
A. Santowski & L. Giese
}
