% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialProjection.R
\name{spatialProjectionion}
\alias{spatialProjectionion}
\alias{spatialProjection}
\title{spatial projection}
\usage{
spatialProjection=function(rst_from,prst_to,outfilepath)
dependent on package "raster" and "rgdal"
}
\arguments{
\item{rst_from}{List of input tif files}

\item{prst_to}{List of designated projected tif files}

\item{outfilepath}{path where output tif files shall be stored}
}
\value{
geo tif data
}
\details{
Projects a tif files into another projection from a given tif file.
}
\examples{
\dontrun{
spatialProjection=function(rst_from,prst_to,outfilepath)
}

}
\author{
A. Santowski & L. Giese
}
