% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalAggregation.R
\name{temporalAggregation}
\alias{temporalAggregation}
\title{Temporal aggregation}
\usage{
temporalAggregation(rstack, rstack_doy, layer_dates, outputfilepathes,
  interval = "fortnight", fun = max, na.rm = TRUE, cores = 4L)
}
\arguments{
\item{rstack}{Raster stack of MODIS composite files}

\item{rstack_doy}{Raster stack of associated day of year information to rstack}

\item{layer_dates}{Date vector corresponding to input layers (see MODIS::temporalComposite)}

\item{outputfilepathes}{Full path and names of all output layers to be written to disk}

\item{interval}{Aggregation interval (see MODIS::temporalComposite),}

\item{fun}{Aggregation function (see MODIS::temporalComposite)}

\item{na.rm}{NA handling (see MODIS::temporalComposite)}

\item{cores}{Paralleliziation (see MODIS::temporalComposite)}
}
\description{
The function is a wrapper arround MODIS::temporalComposit.
It is hardly usefull since it only adds the writing.
}
\examples{
\dontrun{
temporalAggregation(rstack, rstack_doy, layer_dates, outputfilepathes,
interval = "fortnight", fun = max, na.rm = TRUE,
cores = 4L)
}
}
\author{
Santwoski, A. & C. Weber
}
