% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittakerSmoother.R
\name{whittakerSmoother}
\alias{whittakerSmoother}
\title{Whittaker smoother}
\usage{
whittakerSmoother(input, quality_rst, doy_rst, begin, end, whittaker_files, lambda, nIter, threshold)
dependent on package "MODIS"
}
\arguments{
\item{ndvi_oc_rst, }{stacked vegetation index files}

\item{quality_rst, }{stacked quality checked files}

\item{doy_rst, }{stacked day of year files}

\item{lambda, }{integer, strength of smoothing}

\item{nIter, }{integer, number of iteration for the upper envelope fitting}

\item{threshold, }{integer, threshold for outlier values}

\item{prefix, }{character, file name prefix}

\item{suffix, }{character, file name suffix}

\item{whittaker_files, }{path where whittaker output files shall be stored}
}
\value{
geo-tiff data
}
\details{
this function smoothes a time series vegetation index based on satellite data
}
\examples{
\dontrun{
whittakerSmoother(ndvi_oc_rst = ndvi_oc_rst, quality_rst = path_modis_qua_tiles,
                   doy_rst = path_modis_doy_tiles, whittaker_files = path_modis_whittaker_tiles,
                   lambda = 6000, nIter = 3, threshold = 2000
                   prefix = "ab", suffix = "ws")
}
}
\references{
Nauss, T., Detsch, F.
}
\author{
Hanzl, A., Hahn, L.
}
