% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGapsLin.R
\name{fillGapsLin}
\alias{fillGapsLin}
\title{fill gaps by linear interpolation}
\usage{
dependent of package "raster", "foreach", "doParallel", "raster", "rgdal", "GSODTools"
}
\arguments{
\item{raster_stack}{raster stack of scaled and temporal aggregated rasters}

\item{out_path}{path to folder, where output should be stored}
}
\value{
writes out rasters with no NA values
}
\description{
subsets NAs by spacial linear interpolation
}
\examples{
\dontrun{
fill_gaps_lin(raster_stack, outfilepath)
}

}
\references{
Nauss, T., Detsch, F.
}
\author{
Johannes Schnell, Laura Giese
}
