% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalAggregation.R
\name{temporalAggregation}
\alias{temporalAggregation}
\title{Temporal aggregation}
\usage{
temporalAggregation(input)
}
\arguments{
\item{input}{Set input file (default= path_modis_whittaker_tiles, from 00_set_environment call)}
}
\description{
The temporal function takes up the previously created scaled .tif-files (scaled-function).
The function changes the input .tif-files to a stack with temporal composite. The function returns new calculated
.tif-files with pattern /day_of_the_year into the modis_doy_tiles file (The file must be created beforehand).
The function requires the "raster" package.
}
\examples{
\dontrun{
temporalAggregation(input)
}
}
\author{
Santwoski, A. & C. Weber
}
